ajout de adherent dao et php modif ligue dao et modif supprimer club

+++ supprimer_club.php
@@ -53,12 +53,22 @@
 <?php
         if(isset($_POST['Supprimer'])){
           $libmotif = $_POST['libmotif'];
-          $id = $ClubDao->findtheID($libmotif);     
-          $nb = $ClubDao->delete($id);
-          if($nb == 1){
-            echo "<br>Le motif N°$id a bien été Supprimer dans la base FREDI";
+          $id = $ClubDao->findtheID($libmotif);
+          $adhdao = new AdherentDAO;
+          $res = $adhdao->findnb($id);
+          foreach($res as $re)
+            $tot = $re;
+
+          if($tot == 0){
+            $nb = $ClubDao->delete($id);
+            if($nb == 1){
+              echo "<br>Le club $libmotif a bien été Supprimer dans la base FREDI";
+            }else{
+              echo "<br>Le club $libmotif n'a pas été Supprimer dans la base FREDI";
+            }
           }else{
-            echo "<br>Le motif N°$id n'a pas été Supprimer dans la base FREDI";
+            echo "<br>Le club $libmotif ne peut pas être supprimé car au moins un
+            adhérent y est affilié";
           }
         }        
 ?>
